# Hestia-Swagger/swagger.yml
openapi: 3.0.3
info:
  title: Hestia — API Reference
  description: |
    Aggregated API reference for all Hestia services.
    Each path overrides the server URL to route to the correct service.

    | Service   | Port  |
    |-----------|-------|
    | Hub       | 19001 |
    | Archive   | 19002 |
    | Hecate    | 19003 |
    | Oracle    | 19004 |
    | Hermes    | 19005 |
    | Scout     | 19006 |
    | Chronos   | 19007 |
    | Argus     | 19008 |
    | Iris      | 19012 |
    | Hephaestus| 19010 |
    | Dummy     | 19011 |
  version: 1.0.0
servers:
  - url: http://localhost:19001
    description: Hub (default)
...
components:
  schemas:
...
    RegisterServiceRequest:
      type: object
      required:
        [name, base_url, health_endpoint, service_type, service_version, tags]
      properties:
        name:
          type: string
          pattern: "^[a-z0-9_-]{2,40}$"
          example: my_service
        base_url:
          type: string
          example: http://hestia_myservice:9999
        health_endpoint:
          type: string
          example: /health
        service_type:
          type: string
          enum: [core, module, integration]
        service_version:
          type: string
          example: 1.0.0
        tags:
          type: array
          items:
            type: string
            enum:
              [
                core,
                module,
                storage,
                connector,
                chat,
                dispatch,
                real_estate,
                integration,
                messaging,
                monitoring,
              ]
        topology_tags:
          type: array
          items:
            type: string
            pattern: "^(layer|domain|status):[a-z_]+$"
        capabilities:
          type: object
          additionalProperties: true
    RouteRequest:
      type: object
      required: [method]
      properties:
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
        headers:
          type: object
          additionalProperties:
            type: string
        query:
          type: object
          additionalProperties: true
        body:
          type: object
          additionalProperties: true
        timeout_seconds:
          type: number
          default: 8.0

    # ── Archive ───────────────────────────────────────────────────────────────
...
paths:
...
  /api/registry/register:
    post:
      tags: [Hub]
      summary: Register a service
      servers:
        - url: http://localhost:19001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterServiceRequest"
      responses:
        "200":
          description: Registered
  /api/registry/deregister:
    post:
      tags: [Hub]
      summary: Deregister a service
      servers:
        - url: http://localhost:19001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                base_url:
                  type: string
      responses:
        "200":
          description: Deregistered
...
  /api/registry/wait:
    get:
      tags: [Hub]
      summary: Wait for registry revision change (long-poll)
      servers:
        - url: http://localhost:19001
      parameters:
        - name: after_revision
          in: query
          schema:
            type: integer
            default: 0
        - name: timeout_seconds
          in: query
          schema:
            type: number
            default: 120
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  revision:
                    type: integer
                  updated_at:
                    type: number
                  changed:
                    type: boolean
                  services_count:
                    type: integer
...
  /api/discovery/commands:
    get:
      tags: [Hub]
      summary: Discover all routable commands
      servers:
        - url: http://localhost:19001
      parameters:
        - name: client
          in: query
          schema:
            type: string
          example: telegram
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  commands:
                    type: array
                    items:
                      type: object
...
  /api/logs:
    get:
      tags: [Hub]
      summary: Retrieve buffered runtime logs (service-local endpoint)
      description: |
        Implemented in all runtime services.
        Current instances:
        - Hub (19001)
        - Archive (19002)
        - Hecate (19003)
        - Oracle (19004)
        - Hermes (19005)
        - Scout (19006)
        - Chronos (19007)
        - Argus (19008)
        - Athena (19009)
        - Hephaestus (19010)
        - Telegram control API (internal: hestia_telegram:8010)

        Sensitive values (tokens, passwords, secrets, credential URLs, auth headers)
        are redacted before entries are buffered and returned.

        Use query filters to keep output dense and useful for troubleshooting.
      servers:
        - url: http://localhost:19001
          description: Hub
        - url: http://localhost:19002
          description: Archive
        - url: http://localhost:19003
          description: Hecate
        - url: http://localhost:19012
          description: Iris
        - url: http://localhost:19004
          description: Oracle
        - url: http://localhost:19005
          description: Hermes
        - url: http://localhost:19006
          description: Scout
        - url: http://localhost:19007
          description: Chronos
        - url: http://localhost:19008
          description: Argus
        - url: http://localhost:19009
          description: Athena
        - url: http://localhost:19010
          description: Hephaestus
        - url: http://hestia_telegram:8010
          description: Telegram control API (Docker network)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
            maximum: 2000
        - name: level
          in: query
          schema:
            type: string
            example: INFO
        - name: contains
          in: query
          schema:
            type: string
            example: outbound_event_id=p2-4-smoke-3
      responses:
        "200":
          description: "{service, count, logs[]}"
...
  /api/monitor/logs/{service_name}:
    get:
      tags: [Hub]
      summary: Retrieve logs for a registered service through Hub monitoring proxy
      description: |
        Supports multiple output modes:
        - raw: original proxied payload
        - text/transcribe: flattened plain-text log stream
        - ai/analyze/summary: Oracle-formatted analysis, with text fallback
      servers:
        - url: http://localhost:19001
      parameters:
        - name: service_name
          in: path
          required: true
          schema:
            type: string
          example: archive
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
            maximum: 2000
        - name: level
          in: query
          schema:
            type: string
            example: WARNING
        - name: contains
          in: query
          schema:
            type: string
            example: event=hub_register_non_success
        - name: mode
          in: query
          schema:
            type: string
            enum: [raw, text, transcribe, ai, analyze, summary]
            default: raw
        - name: response_prompt
          in: query
          schema:
            type: string
            example: Summarize critical errors and actionable next checks.
        - name: include_raw
          in: query
          schema:
            type: boolean
            default: false
        - name: timeout_seconds
          in: query
          schema:
            type: number
            default: 8.0
      responses:
        "200":
          description: "Mode-dependent payload. Raw: {status_code, service, target, payload}; Text/AI: {status_code, service, target, mode, text, line_count[, payload]}"
  /api/route/{service_name}/{path}:
    parameters:
      - name: service_name
        in: path
        required: true
        schema:
          type: string
        description: |
          Service name, comma-separated list, or ``*`` wildcard.

          | Value | Mode | Example |
          |-------|------|---------|
          | ``oracle`` | Unicast | Route to oracle only |
          | ``oracle,argus`` | Multicast | Route to oracle AND argus |
          | ``*`` | Broadcast | Route to ALL registered services |

          Broadcast/multicast returns an aggregated envelope:
          ``{"service": "*", "count": N, "ok": N, "results": [...]}``.
          Unicast returns the target's payload directly (backward-compatible).
      - name: path
        in: path
        required: true
        schema:
          type: string
        example: api/logs/level
    get:
      tags: [Hub]
      summary: Route a GET request (unicast, multicast, or broadcast)
      description: |
        Use ``service_name=*`` to broadcast to all services.
        Use ``service_name=oracle,argus`` for multicast.
      servers:
        - url: http://localhost:19001
      parameters:
        - name: service_name
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Aggregated results (multicast/broadcast) or direct payload (unicast)
    post:
      tags: [Hub]
      summary: Route a request to a registered service (unicast, multicast, or broadcast)
      description: |
        Use ``service_name=*`` to broadcast to all services (e.g. set all log levels at once).
        Use ``service_name=oracle,argus`` for multicast to specific services.
      servers:
        - url: http://localhost:19001
      parameters:
        - name: service_name
          in: path
          required: true
          schema:
            type: string
          example: archive
        - name: path
          in: path
          required: true
          schema:
            type: string
          example: api/entities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteRequest"
      responses:
        "200":
          description: Proxied response from the target service, or aggregated envelope for broadcast/multicast
...